+++ NEWS.md
@@ -17,6 +17,7 @@
 * Added `as_html()` for obtaining table as `htmltools` tags.
 * `to_screen()` now displays double, dashed and dotted border styles.
 * Bugfix: Typst output now omits borders when none are set.
+* Bugfix: Typst captions now escape special characters.
 
 
 # huxtable 5.6.0